refactor(logs): move Docker log parsing into its own package
