+++ internal/logs/logs.go
@@ -1,4 +1,4 @@
-package api
+package logs
 
 import (
 	"context"
