+++ internal/logs/logs_test.go
@@ -1,4 +1,4 @@
-package api
+package logs
 
 import (
 	"bytes"
